warnings

+++ userspace/libsinsp/utils.cpp
@@ -1037,54 +1037,54 @@ void replace_in_place(string& str, string& substr_to_replace, string& new_substr
 ///////////////////////////////////////////////////////////////////////////////
 uint8_t sinsp_numparser::parseu8(const string& str)
 {
-	uint8_t res;
+	uint32_t res;
 	char temp;
 
-	if(std::sscanf(str.c_str(), "%" PRIu8 "%c", &res, &temp) != 1)
+	if(std::sscanf(str.c_str(), "%" PRIu32 "%c", &res, &temp) != 1)
 	{
 		throw sinsp_exception(str + " is not a valid number");
 	}
 
-	return res;
+	return (uint8_t)res;
 }
 
 int8_t sinsp_numparser::parsed8(const string& str)
 {
-	int8_t res;
+	int32_t res;
 	char temp;
 
-	if(std::sscanf(str.c_str(), "%" PRId8 "%c", &res, &temp) != 1)
+	if(std::sscanf(str.c_str(), "%" PRId32 "%c", &res, &temp) != 1)
 	{
 		throw sinsp_exception(str + " is not a valid number");
 	}
 
-	return res;
+	return (int8_t)res;
 }
 
 uint16_t sinsp_numparser::parseu16(const string& str)
 {
-	uint16_t res;
+	uint32_t res;
 	char temp;
 
-	if(std::sscanf(str.c_str(), "%" PRIu16 "%c", &res, &temp) != 1)
+	if(std::sscanf(str.c_str(), "%" PRIu32 "%c", &res, &temp) != 1)
 	{
 		throw sinsp_exception(str + " is not a valid number");
 	}
 
-	return res;
+	return (uint16_t)res;
 }
 
 int16_t sinsp_numparser::parsed16(const string& str)
 {
-	int16_t res;
+	int32_t res;
 	char temp;
 
-	if(std::sscanf(str.c_str(), "%" PRId16 "%c", &res, &temp) != 1)
+	if(std::sscanf(str.c_str(), "%" PRId32 "%c", &res, &temp) != 1)
 	{
 		throw sinsp_exception(str + " is not a valid number");
 	}
 
-	return res;
+	return (int16_t)res;
 }
 
 uint32_t sinsp_numparser::parseu32(const string& str)
